Fix rgb image plot
- Fix the plotting of rgb images. plt.imshow was plotting rgb images as black and white

+++ gp_retouch/image/image.py
@@ -67,7 +67,7 @@ class Image:
 
     def plot(self):  # noqa: D102
         if self.is_rgb:
-            plt.imshow(self.data)
+            plt.imshow(self.data.astype(np.uint8))
         else:
             plt.imshow(self.data, cmap='gray', vmin=0, vmax=255)
         plt.axis("off")
